requirements.txt
fastapi~=0.111.0
pydantic~=2.8.2
uvicorn~=0.30.1
python-dotenv~=1.0.1
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
pymongo~=4.8.0
openai~=1.35.10
transformers~=4.42.3
pillow~=10.4.0
filetype~=1.2.0